refactor: enhance product and product price item
