+++ src/main/kotlin/org/shangahi/sellio_backend/api/swagger/doc/ProductDoc.kt
@@ -7,7 +7,7 @@ import io.swagger.v3.oas.annotations.media.Schema
 import io.swagger.v3.oas.annotations.parameters.RequestBody
 import io.swagger.v3.oas.annotations.responses.ApiResponse
 import org.shangahi.sellio_backend.api.dto.request.ProductRequest
-import org.shangahi.sellio_backend.api.dto.request.ProductResponse
+import org.shangahi.sellio_backend.api.dto.response.ProductResponse
 import org.shangahi.sellio_backend.api.dto.response.ErrorResponse
 import org.shangahi.sellio_backend.api.swagger.ErrorResponseExample
